Restore overscope parent on exit
