+++ R/eval-tidy.R
@@ -319,6 +319,6 @@ overscope_clean <- function(overscope) {
 
 f_self_eval <- function(overscope, overscope_top) {
   function(lhs, rhs) {
-    .Call(rlang_tilde_eval, sys.call(), overscope, overscope_top)
+    .Call(rlang_tilde_eval, sys.call(), overscope, overscope_top, environment())
   }
 }
